requirements.txt
numpy
pyyaml
matplotlib
open3d
torch
